+++ src/ui/ProcessTab.py
@@ -498,6 +498,8 @@ class ProcessTab:
             f"{self.settings.settings['pytorch_gpu_id']}",
             "--cwd",
             f"{CWD}",
+            "--override_upscale_scale",
+            f"{renderOptions.overrideUpscaleScale}",
         ]
         if renderOptions.startTime is not None:
             command += [