+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Lock to make sure the app does not have duplicates open.
  - Switchable pytorch backend and version.
  - Experimental Blackwell GPU support. 
+ - Selectable upscale scale independent of model scale.
 ### Changed
  - Installes to a preset directory, and updates python and backend dynamically.
  - Improved RIFE TensorRT speeds